# OPENAPI/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paperless-AI API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Paperless-AI application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Clusterzx",
      "url": "https://github.com/Clusterzx"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT-based authentication for web app users. The token is obtained by authenticating via the login endpoint.\n\n### How to authenticate:\n1. Send a POST request to `/login` with your username and password\n2. The server will respond with a JWT token (also set as a cookie in browsers)\n3. Include this token in the `Authorization` header as `Bearer {token}`\n\nExample:\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nJWT tokens are valid for 24 hours after issuance.\n"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key authentication for programmatic access. The API key can be generated or regenerated using the /api/key-regenerate endpoint.\n\n### How to authenticate:\n1. Access the API key from your application settings\n2. Include the API key in the `x-api-key` HTTP header for all requests\n\nExample:\n```\nx-api-key: 7c1f3f5e2b0a9d8c6e4b2a1d3f5e8c9b2a1d3f5e\n```\n\nAPI keys do not expire unless regenerated.\n"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Error resetting documents"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/api-docs/openapi.json": {
      "get": {
        "summary": "Retrieve the OpenAPI specification",
        "description": "Returns the complete OpenAPI specification for the Paperless-AI API.\nThis endpoint attempts to serve a static OpenAPI JSON file first, falling back\nto dynamically generating the specification if the file cannot be read.\n\nThe OpenAPI specification document contains all API endpoints, parameters,\nrequest bodies, responses, and schemas for the entire application.\n",
        "tags": [
          "API",
          "System"
        ],
        "responses": {
          "200": {
            "description": "OpenAPI specification returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The complete OpenAPI specification"
                }
              }
            }
          },
          "404": {
            "description": "OpenAPI specification file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error occurred while retrieving the OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tags/merge": {
      "post": {
        "summary": "Merge similar tags",
        "description": "Groups similar tags based on name similarity and merges them,\nkeeping the tag with the highest document count as the target.\n",
        "tags": [
          "Tags"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "similarity_threshold",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 1,
              "default": 0.8
            },
            "description": "Threshold for considering tags similar (0.0 to 1.0)"
          }
        ],
        "responses": {
          "200": {
            "description": "Tags merged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mergeCount": {
                      "type": "number",
                      "description": "Number of tags that were merged"
                    },
                    "mergeDetails": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Details of each merge operation"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while merging tags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "summary": "Main dashboard page",
        "description": "Renders the main dashboard page of the application with summary statistics and visualizations.\nThe dashboard provides an overview of processed documents, system metrics, and important statistics\nabout document processing including tag counts, correspondent counts, and token usage.\n\nThe page displays visualizations for document processing status, token distribution, \nprocessing time statistics, and document type categorization to help administrators\nunderstand system performance and document processing patterns.\n",
        "tags": [
          "Navigation",
          "System"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard page rendered successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "HTML content of the dashboard page"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user not authenticated",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/login"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/summaries": {
      "get": {
        "summary": "Render the document summaries tool page",
        "description": "Provides a UI for generating and managing AI summaries of documents.\n",
        "tags": [
          "Navigation",
          "Documents"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Document summaries tool page rendered successfully"
          },
          "401": {
            "description": "Unauthorized - user not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
...
}